+++ Editor/Scripts/VisualScriptingSetupEditor.cs
@@ -1,12 +1,12 @@
 #if UNITY_EDITOR
+using Reflectis.SDK.Core.Editor;
+
 using System;
 using System.Collections.Generic;
 
 using Unity.VisualScripting;
 
 using UnityEditor;
-
-using UnityEngine;
 
 namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
 {
@@ -20,7 +20,7 @@ namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
                 BoltCore.Configuration.typeOptions.RemoveAt(0);
             }
 
-            var _typeOptionsMetadata = BoltCore.Configuration.GetMetadata(nameof(BoltCore.Configuration.typeOptions));
+            PluginConfigurationItemMetadata _typeOptionsMetadata = BoltCore.Configuration.GetMetadata(nameof(BoltCore.Configuration.typeOptions));
 
             if (_typeOptionsMetadata.defaultValue is List<Type> defaultTypes)
             {
@@ -31,7 +31,7 @@ namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
             }
         }
 
-        [MenuItem("Reflectis/Setup Visual Scripting Nodes")]
+        [MenuItem("Reflectis Worlds/Creator Kit/Visual Scripting/Setup Visual Scripting Nodes")]
         public static void Setup()
         {
             if (!VSUsageUtility.isVisualScriptingUsed)
@@ -45,7 +45,7 @@ namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
             {
                 if (type == null)
                 {
-                    Debug.LogError("Found Null type!");
+                    //Debug.LogError("Found Null type!");
                     continue;
                 }
                 if (!BoltCore.Configuration.typeOptions.Contains(type))
@@ -59,20 +59,18 @@ namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
                     BoltCore.Configuration.typeOptions.Add(type);
                     //Debug.Log($"Add {type.FullName} to Visual Scripting type options.");
                 }
-
             }
-
             var typeOptionsMetadata = BoltCore.Configuration.GetMetadata(nameof(BoltCore.Configuration.typeOptions));
             typeOptionsMetadata.Save();
-            BoltCore.Configuration.Save();
+            BoltCore.Configuration.SaveProjectSettingsAsset(true);
             Codebase.UpdateSettings();
             UnitBase.Rebuild();
         }
 
         private static IEnumerable<Type> GetCustomTypes()
         {
-            VisualScriptingCustomTypeCollector[] customTypeCollectors = Resources.FindObjectsOfTypeAll<VisualScriptingCustomTypeCollector>();
-
+            UnityEditor.AssetDatabase.Refresh();
+            List<VisualScriptingCustomTypeCollector> customTypeCollectors = AssetDatabaseExtension.SearchAssetByType<VisualScriptingCustomTypeCollector>();
             List<Type> customTypes = new();
             foreach (var customTypeCollector in customTypeCollectors)
             {
@@ -80,6 +78,10 @@ namespace Reflectis.CreatorKit.Worlds.VisualScripting.Editor
             }
             return customTypes;
         }
+
+
     }
+
+
 }
 #endif